src/requirements.txt
flask>=2.2.5
gunicorn>=21.2.0
python-dotenv>=1.0.0
requests>=2.31.0
